fix line breaks
If I were to guess, this might be coming from line breaks being inserted when someone creates a pull request through the cool 'edit this lesson' gizmo.

Just a guess.

It's also possible that a developer has something weird set up in their text editor.  It's not every lesson, and it's not every paragraph (sometimes it is though).